+++ CHANGES.md
@@ -2,7 +2,8 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
-*
+* Add options.maxBufferSize and optinons.bufferFlushInterval
+* Change options.global_tags to options.globalTags for conistency
 
 --------------------
 
@@ -10,5 +11,5 @@ CHANGELOG
 * Thrown error when cacheDNS flag fails to resolve DNS name
 
 ## 1.0.1 (2015-09-24)
-* Start from the base of https://github.com/sivy/node-statsd
 * Add the event API used by DogStatsD
+* Start from the base of https://github.com/sivy/node-statsd